actually working lines

+++ Point.cpp
@@ -1,16 +1,16 @@
 #include "Point.h"
 
-Point::Point(int x, int y, size_t thickness) {
+Point::Point(float x, float y, size_t thickness) {
   this->x = x;
   this->y = y;
   this->h = thickness;
   this->w = thickness;
 }
-Point::Point(int x, int y) {
+Point::Point(float x, float y) {
   this->x = x;
   this->y = y;
   this->h = 5;
   this->w = 5;
 }
-int Point::getX() { return x; }
-int Point::getY() { return y; }
+float Point::getX() { return x; }
+float Point::getY() { return y; }
